backend/app/requirements.txt
fastapi
uvicorn
celery
redis
sqlalchemy
httpx
python-dotenv
pydantic
email-validator